+++ lib/b_views/h_app_settings/a_app_settings_screen/a_app_settings_screen.dart
@@ -42,7 +42,7 @@ class AppSettingsScreen extends StatelessWidget {
         /// TEST
         const PyramidFloatingButton(
           icon: Iconz.lab,
-          color: Colorz.red255,
+          color: Colorz.yellow255,
           onTap: superDevTest,
         ),