+++ lib/b_views/j_flyer/z_components/c_groups/grid/components/flyers_grid_builder.dart
@@ -1,3 +1,4 @@
+import 'package:basics/animators/helpers/app_scroll_behavior.dart';
 import 'package:basics/bldrs_theme/classes/ratioz.dart';
 import 'package:bldrs/b_views/j_flyer/z_components/x_helpers/x_flyer_dim.dart';
 import 'package:flutter/material.dart';
@@ -10,8 +11,8 @@ class FlyersGridBuilder extends StatelessWidget {
     required this.builder,
     required this.itemCount,
     required this.hasResponsiveSideMargin,
+    required this.numberOfColumnsOrRows,
     this.topPadding = Ratioz.stratosphere,
-    this.numberOfColumnsOrRows = 2,
     this.scrollDirection = Axis.vertical,
     this.scrollable = true,
     this.scrollController,
@@ -57,28 +58,31 @@ class FlyersGridBuilder extends StatelessWidget {
       key: const ValueKey<String>('FlyersGridBuilder'),
       width: _gridWidth,
       height: _gridHeight,
-      child: GridView.builder(
-          controller: scrollController,
-          physics: scrollable == true ? const BouncingScrollPhysics() : const NeverScrollableScrollPhysics(),
-          padding: FlyerDim.flyerGridPadding(
-            gridWidth: _gridWidth,
-            gridHeight: _gridHeight,
-            context: context,
-            topPaddingValue: topPadding,
-            endPadding: bottomPadding,
-            gridSpacingValue: FlyerDim.flyerGridGridSpacingValue(_gridSlotWidth),
-            isVertical: scrollDirection == Axis.vertical,
-            hasResponsiveSideMargin: hasResponsiveSideMargin,
-          ),
-          gridDelegate: FlyerDim.flyerGridDelegate(
-            context: context,
-            flyerBoxWidth: _gridSlotWidth,
-            numberOfColumnsOrRows: numberOfColumnsOrRows,
+      child: ScrollConfiguration(
+        behavior: const AppScrollBehavior(),
+        child: GridView.builder(
+            controller: scrollController,
+            physics: scrollable == true ? const BouncingScrollPhysics() : const NeverScrollableScrollPhysics(),
+            padding: FlyerDim.flyerGridPadding(
+              gridWidth: _gridWidth,
+              gridHeight: _gridHeight,
+              context: context,
+              topPaddingValue: topPadding,
+              endPadding: bottomPadding,
+              gridSpacingValue: FlyerDim.flyerGridGridSpacingValue(_gridSlotWidth),
+              isVertical: scrollDirection == Axis.vertical,
+              hasResponsiveSideMargin: hasResponsiveSideMargin,
+            ),
+            gridDelegate: FlyerDim.flyerGridDelegate(
+              context: context,
+              flyerBoxWidth: _gridSlotWidth,
+              numberOfColumnsOrRows: numberOfColumnsOrRows,
+              scrollDirection: scrollDirection,
+            ),
             scrollDirection: scrollDirection,
-          ),
-          scrollDirection: scrollDirection,
-          itemCount: itemCount,
-          itemBuilder: builder,
+            itemCount: itemCount,
+            itemBuilder: builder,
+        ),
       ),
     );