+++ lib/b_views/j_flyer/z_components/d_variants/a_flyer_box.dart
@@ -1,9 +1,9 @@
 import 'package:basics/bldrs_theme/classes/colorz.dart';
-import 'package:basics/super_box/src/f_super_box_tap_layer/x_tap_layer.dart';
+import 'package:basics/bldrs_theme/classes/shadowers.dart';
 import 'package:bldrs/b_views/j_flyer/z_components/x_helpers/x_flyer_dim.dart';
 import 'package:flutter/material.dart';
 
-/// VARS OPTIMIZED
+/// OLD PERFECT
 class FlyerBox extends StatelessWidget {
   /// --------------------------------------------------------------------------
   const FlyerBox({
@@ -34,20 +34,26 @@ class FlyerBox extends StatelessWidget {
     );
     // --------------------
     return Center(
-      key: const ValueKey<String>('flyerBoxWidget'),
-      child: TapLayer(
+      key: const ValueKey<String>('flyer_box'),
+      child: InkWell(
         onTap: onTap,
         splashColor: boxColor?.withOpacity(0.2),
-        corners: _flyerBorders,
-        width: flyerBoxWidth,
-        height: _flyerBoxHeight,
-        boxColor: boxColor,
-        alignment: Alignment.topCenter,
-        child: ClipRRect(
-          borderRadius: FlyerDim.flyerCorners(flyerBoxWidth),
-          child: Stack(
-            alignment: Alignment.topCenter,
-            children: stackWidgets ?? <Widget>[],
+        borderRadius: _flyerBorders,
+        child: Container(
+          width: flyerBoxWidth,
+          height: _flyerBoxHeight,
+          alignment: Alignment.topCenter,
+          decoration: BoxDecoration(
+            color: boxColor,
+            borderRadius: _flyerBorders,
+            boxShadow: shadowIsOn == true ? Shadower.flyerZoneShadow : null,
+          ),
+          child: ClipRRect(
+            borderRadius: _flyerBorders,
+            child: Stack(
+              alignment: Alignment.topCenter,
+              children: stackWidgets ?? <Widget>[],
+            ),
           ),
         ),
       ),
@@ -56,3 +62,60 @@ class FlyerBox extends StatelessWidget {
   }
   // -----------------------------------------------------------------------------
 }
+
+/// NEW PROPOSAL
+// class FlyerBox extends StatelessWidget {
+//   /// --------------------------------------------------------------------------
+//   const FlyerBox({
+//     required this.flyerBoxWidth,
+//     this.stackWidgets,
+//     this.boxColor = Colorz.white20,
+//     this.onTap,
+//     this.shadowIsOn = false,
+//     super.key
+//   });
+//   /// --------------------------------------------------------------------------
+//   // --- NEAT AND CLEAN
+//   /// width factor * screenWidth = flyerWidth
+//   final double flyerBoxWidth;
+//   /// internal parts of the flyer
+//   final List<Widget>? stackWidgets;
+//   final Color? boxColor;
+//   final void Function()? onTap;
+//   final bool shadowIsOn;
+//   // -----------------------------------------------------------------------------
+//   @override
+//   Widget build(BuildContext context) {
+//     // --------------------
+//     final BorderRadius _flyerBorders = FlyerDim.flyerCorners(flyerBoxWidth);
+//     // --------------------
+//     final double _flyerBoxHeight = FlyerDim.flyerHeightByFlyerWidth(
+//       flyerBoxWidth: flyerBoxWidth,
+//     );
+//     // --------------------
+//     return Center(
+//       key: const ValueKey<String>('flyerBoxWidget'),
+//       child: AspectRatio(
+//         aspectRatio: FlyerDim.flyerAspectRatio(),
+//         child: TapLayer(
+//           onTap: onTap,
+//           splashColor: boxColor?.withOpacity(0.2),
+//           corners: _flyerBorders,
+//           width: flyerBoxWidth,
+//           height: _flyerBoxHeight,
+//           boxColor: boxColor,
+//           alignment: Alignment.topCenter,
+//           child: ClipRRect(
+//             borderRadius: FlyerDim.flyerCorners(flyerBoxWidth),
+//             child: Stack(
+//               alignment: Alignment.topCenter,
+//               children: stackWidgets ?? <Widget>[],
+//             ),
+//           ),
+//         ),
+//       ),
+//     );
+//     // --------------------
+//   }
+//   // -----------------------------------------------------------------------------
+// }